refactor: unsecret env in prerequest
